refactor: move api related utils into api/utils
