+++ src/api/storybook.index.js
@@ -1,6 +1,7 @@
-/* eslint-disable import/prefer-default-export */
 import { sleep } from '@noaignite/utils'
 import { filter, products } from '~/api/__mock__'
+
+export * from './utils'
 
 export async function fetchProducts(query) {
   await sleep(500)
