+++ src/api/utils/linkType.js
