+++ src/containers/Cart/Cart.js
@@ -2,8 +2,8 @@ import * as React from 'react'
 import PropTypes from 'prop-types'
 import { styled } from '@mui/system'
 import { Typography } from '@mui/material'
+import { cartItemType } from '~/api'
 import { useCentraSelection, useI18n } from '~/context'
-import { cartItemType } from '~/utils'
 import CartItem from '../CartItem'
 import CartSummary from '../CartSummary'
