+++ src/containers/CartSummary/CartSummary.js
@@ -2,8 +2,8 @@ import * as React from 'react'
 import PropTypes from 'prop-types'
 import { styled } from '@mui/system'
 import { Typography } from '@mui/material'
+import { cartItemType } from '~/api'
 import { useCentraSelection, useI18n } from '~/context'
-import { cartItemType } from '~/utils'
 
 const CartSummaryRoot = styled('div', {
   name: 'CartSummary',
