+++ src/utils/index.js
@@ -1,5 +1,3 @@
-export { default as cartItemType } from './cartItemType'
+export { default as createEmotionCache } from './createEmotionCache'
 export { default as createRenderBlock } from './createRenderBlock'
-export { default as linkType } from './linkType'
-export { default as menuLinkType } from './menuLinkType'
-export { default as productType } from './productType'
+export { default as getConfig } from './getConfig'
